ref: extracted a base request from Game- & PredictionRequest
