+++ src/main/java/com/aau/wizard/dto/request/PredictionRequest.java
@@ -1,32 +1,16 @@
 package com.aau.wizard.dto.request;
 
-public class PredictionRequest {
-    private String gameId;
-    private String playerId;
+public class PredictionRequest extends BaseRequest {
     private int prediction;
 
+    /**
+     * No args constructor needed for Jackson / JSON deserialization
+     */
     public PredictionRequest(){}
 
     public PredictionRequest(String gameId, String playerId, int prediction){
-        this.gameId=gameId;
-        this.playerId=playerId;
-        this.prediction=prediction;
-    }
-
-    public String getGameId() {
-        return gameId;
-    }
-
-    public void setGameId(String gameId) {
-        this.gameId = gameId;
-    }
-
-    public String getPlayerId() {
-        return playerId;
-    }
-
-    public void setPlayerId(String playerId) {
-        this.playerId = playerId;
+        super(gameId, playerId);
+        this.prediction = prediction;
     }
 
     public int getPrediction() {
